+++ src/components/interactive/DropdownButton.jsx
@@ -3,10 +3,12 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 import CaretDownIcon from 'lib/icons/react/CaretDownIcon';
 
-const DropdownButton = ({ children, ...rest }) => (
+const DropdownButton = ({ children, isFullWidth = false, ...rest }) => (
   <button
     {...rest}
-    className={cx('display--inlineFlex', 'fdsDropdown', 'border--focus', 'rounded--all')}
+    className={cx('display--inlineFlex', 'fdsDropdown', 'border--focus', 'rounded--all', {
+      'fdsButton--isFullWidth': isFullWidth,
+    })}
   >
     <div>{children}</div>
     <div className="margin--left--half alignChild--center--center">
@@ -20,6 +22,8 @@ DropdownButton.propTypes = {
    * content rendered to the left of the carat
    */
   children: PropTypes.node,
+  /** Controls the button going full width */
+  isFullWidth: PropTypes.bool,
 };
 
 export default DropdownButton;