+++ src/components/modals/Dialog.jsx
@@ -18,12 +18,23 @@ export const isElementOverflowing = ({ current }) => {
   return false;
 };
 
-const Dialog = (props) => {
+const Dialog = ({
+  role = 'dialog',
+  width = '500px',
+  height = '80vh',
+  disableFocusTrap = false,
+  onDismiss,
+  alwaysShowBorder,
+  isOpen,
+  title,
+  content,
+  footerContent,
+}) => {
   const contentEl = useRef(null);
 
   const handleKeyDown = (e) => {
-    if (props.onDismiss && e.key === 'Escape') {
-      props.onDismiss();
+    if (onDismiss && e.key === 'Escape') {
+      onDismiss();
     }
   };
 
@@ -34,87 +45,92 @@ const Dialog = (props) => {
   };
 
   useEffect(() => {
-    handleResize(); // needs to fire once immediately on mount
-    // eslint-disable-next-line no-undef
-    window.addEventListener('resize', handleResize);
-    return () => {
-      noScroll.off();
+    if (!alwaysShowBorder) {
+      handleResize(); // needs to fire once immediately on mount
       // eslint-disable-next-line no-undef
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
+      window.addEventListener('resize', handleResize);
+      return () => {
+        noScroll.off();
+        // eslint-disable-next-line no-undef
+        window.removeEventListener('resize', handleResize);
+      };
+    }
+    return undefined;
+  }, [alwaysShowBorder]);
 
   useLayoutEffect(() => {
     // This toggles scrolling on and off based on whether the modal
     // is shown or not
-    if (props.isOpen) {
+    if (isOpen) {
       noScroll.on();
     } else {
       noScroll.off();
     }
-  }, [props.isOpen]);
+  }, [isOpen]);
+
+  const dialogNode = (
+    <div>
+      <div className="dialog-overlay"></div>
+      <div className="dialog-zIndex dialog-wrapper">
+        <div
+          className="dialog elevation--3"
+          role={role}
+          aria-labelledby={title && 'a11y-dialog-title'}
+          aria-describedby="a11y-dialog-desc"
+          tabIndex="-1"
+          aria-modal="true"
+          onKeyDown={handleKeyDown}
+          style={{
+            maxWidth: `${width}${typeof width === 'number' ? 'px' : ''} `,
+            maxHeight: `${height}${typeof height === 'number' ? 'px' : ''}`,
+          }}
+        >
+          {(title || onDismiss) && (
+            <React.Fragment>
+              <div className="dialog-header">
+                <Section border="bottom">
+                  <div className="padding--right--double type--head4">
+                    {title ? <span id="a11y-dialog-title">{title}</span> : '\u00A0'}{' '}
+                    {/* There always needs to be something (even a space) in the header for display reasons */}
+                  </div>
+                  {onDismiss && (
+                    <div className="dialog-icon">
+                      <IconButton
+                        Icon={DenyIcon}
+                        onClick={onDismiss}
+                        aria-label="close"
+                        label="close"
+                      />
+                    </div>
+                  )}
+                </Section>
+              </div>
+            </React.Fragment>
+          )}
+          <div className="dialog-content" ref={contentEl}>
+            <Section>{content}</Section>
+          </div>
+          {footerContent && (
+            <div className="dialog-footer">
+              <Section
+                border={alwaysShowBorder || isOverflowing ? 'top' : undefined}
+                bgColor="white"
+              >
+                {footerContent}
+              </Section>
+            </div>
+          )}
+        </div>
+      </div>
+    </div>
+  );
 
   return (
     <React.Fragment>
       {ReactDOM.createPortal(
-        <CSSTransition timeout={200} in={props.isOpen} classNames="dialog" unmountOnExit>
+        <CSSTransition timeout={200} in={isOpen} classNames="dialog" unmountOnExit>
           <React.Fragment>
-            <FocusTrap>
-              <div>
-                <div className="dialog-overlay"></div>
-                <div className="dialog-zIndex dialog-wrapper">
-                  <div
-                    className="dialog elevation--3"
-                    role={props.role}
-                    aria-labelledby={props.title && 'a11y-dialog-title'}
-                    aria-describedby="a11y-dialog-desc"
-                    tabIndex="-1"
-                    aria-modal="true"
-                    onKeyDown={handleKeyDown}
-                    style={{ maxWidth: `${props.width}px` }}
-                  >
-                    {(props.title || props.onDismiss) && (
-                      <React.Fragment>
-                        <div className="dialog-header">
-                          <Section border="bottom">
-                            <div className="padding--right--double type--head4">
-                              {props.title ? (
-                                <span id="a11y-dialog-title">{props.title}</span>
-                              ) : (
-                                '\u00A0'
-                              )}{' '}
-                              {/* There always needs to be something (even a space) in the header for display reasons */}
-                            </div>
-                            {props.onDismiss && (
-                              <div className="dialog-icon">
-                                <IconButton
-                                  Icon={DenyIcon}
-                                  onClick={props.onDismiss}
-                                  aria-label="close"
-                                />
-                              </div>
-                            )}
-                          </Section>
-                        </div>
-                      </React.Fragment>
-                    )}
-                    <div className="dialog-content" ref={contentEl}>
-                      <Section>{props.content}</Section>
-                    </div>
-                    {props.footerContent && (
-                      <div className="dialog-footer">
-                        <Section
-                          border={isOverflowing ? 'top' : undefined}
-                          bgColor="white"
-                        >
-                          {props.footerContent}
-                        </Section>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </FocusTrap>
+            {disableFocusTrap ? dialogNode : <FocusTrap>{dialogNode}</FocusTrap>}
           </React.Fragment>
         </CSSTransition>,
         // eslint-disable-next-line no-undef
@@ -122,11 +138,6 @@ const Dialog = (props) => {
       )}
     </React.Fragment>
   );
-};
-
-Dialog.defaultProps = {
-  role: 'dialog',
-  width: 500,
 };
 
 Dialog.propTypes = {
@@ -152,7 +163,22 @@ Dialog.propTypes = {
   onDismiss: PropTypes.func,
 
   /** Custom modal width */
-  width: PropTypes.number,
+  width: PropTypes.string,
+
+  /** Custom modal height */
+  height: PropTypes.string,
+
+  /**
+   * Border between content area and footer currently renders only if there is scrollable content. Set this boolean
+   * so that the border always renders no matter what.
+   */
+  alwaysShowBorder: PropTypes.bool,
+
+  /**
+   * Disables the focus trap on the Dialog.
+   * Useful when the Dialog contains components that manage focus (e.g. `Menu`)
+   */
+  disableFocusTrap: PropTypes.bool,
 };
 
 export default Dialog;