+++ src/components/util/storybook.js
@@ -1,10 +1,8 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
+import CloneIcon from 'lib/icons/react/CloneIcon';
+import CheckIcon from 'lib/icons/react/CheckIcon';
 import FDS from '../../../lib/dictionary/js/styleConstants';
-
-/**
- * Storybook helpers
- */
 
 /**
  * Creates a knobs options object where labels match values.
@@ -21,41 +19,21 @@ export const arrayToOptions = (values) =>
     {}
   );
 
-/**
- * @Object
- * useful background images for stories
- */
-export const storyBackgrounds = {
-  diagonalLine:
-    'url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAJ0lEQVQYV2NkQAP/b///z4gsBhZQZWSEC8IEQIrAgsgCYEF0AZAgAAvKE968p7/mAAAAAElFTkSuQmCC)',
-  check:
-    'url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAcAAAAHCAYAAADEUlfTAAAAHElEQVQYV2M8fPiwLwMOwAiStLW13YxNftBJAgAx2BqeI9XcBAAAAABJRU5ErkJggg==)',
+export const useClipboard = () => {
+  const [copiedText, setCopiedText] = useState('');
+  const copyToClipboard = (value) => {
+    // eslint-disable-next-line no-undef
+    window.navigator.clipboard.writeText(value);
+    setCopiedText(value);
+    setTimeout(() => setCopiedText(''), 2000);
+  };
+  return [copiedText, copyToClipboard];
 };
 
-export const StoryWrapper = ({ children }) => (
-  <div
-    className="display--inlineFlex"
-    style={{
-      flexWrap: 'wrap',
-      marginBottom: `-${FDS.SPACE_HALF}`,
-      marginRight: `-${FDS.SPACE_HALF}`,
-      alignItems: 'center',
-    }}
-  >
-    {children}
-  </div>
-);
-
-StoryWrapper.propTypes = {
-  children: PropTypes.node,
-};
-
-export const StoryItem = ({ children }) => (
-  <div className="margin--right--half margin--bottom--half">{children}</div>
-);
-
-StoryItem.propTypes = {
-  children: PropTypes.node,
+export const useToggleHover = () => {
+  const [isHovered, setIsHovered] = useState(false);
+  const toggleHover = () => setIsHovered(!isHovered);
+  return [isHovered, toggleHover];
 };
 
 /**
@@ -97,6 +75,79 @@ Info.propTypes = {
   children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node])
     .isRequired,
   type: PropTypes.oneOf(['warn', 'info']),
+};
+
+export const StoryWrapper = ({ children }) => (
+  <div
+    className="display--inlineFlex"
+    style={{
+      flexWrap: 'wrap',
+      marginBottom: `-${FDS.SPACE_HALF}`,
+      marginRight: `-${FDS.SPACE_HALF}`,
+      alignItems: 'center',
+    }}
+  >
+    {children}
+  </div>
+);
+
+StoryWrapper.propTypes = {
+  children: PropTypes.node,
+};
+
+export const StoryItem = ({ children }) => (
+  <div className="margin--right--half margin--bottom--half">{children}</div>
+);
+
+StoryItem.propTypes = {
+  children: PropTypes.node,
+};
+
+export const storyBackgrounds = {
+  diagonalLine:
+    'url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAJ0lEQVQYV2NkQAP/b///z4gsBhZQZWSEC8IEQIrAgsgCYEF0AZAgAAvKE968p7/mAAAAAElFTkSuQmCC)',
+  check:
+    'url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAcAAAAHCAYAAADEUlfTAAAAHElEQVQYV2M8fPiwLwMOwAiStLW13YxNftBJAgAx2BqeI9XcBAAAAABJRU5ErkJggg==)',
+};
+
+export const ImportPath = ({ component, section }) => {
+  const toggleHover = useToggleHover()[1];
+  const [copiedText, copyToClipboard] = useClipboard();
+
+  const path = `import ${component.displayName} from '@cbinsights/fds/lib/components/${section}/${component.displayName}'`;
+
+  const pathParts = path.split(' ');
+
+  return (
+    <p
+      className="importPath margin--bottom"
+      onMouseEnter={toggleHover}
+      onMouseLeave={toggleHover}
+      onClick={() => copyToClipboard(path)}
+    >
+      <span>{pathParts[0]} </span>
+      <span className="importPath--highlight">{pathParts[1]} </span>
+      <span>{pathParts[2]} </span>
+      <span className="importPath--highlight">{pathParts[3]}</span>
+      {!copiedText ? (
+        <CloneIcon customSize={14} />
+      ) : (
+        <CheckIcon customSize={14} color="green" />
+      )}
+    </p>
+  );
+};
+
+ImportPath.propTypes = {
+  component: PropTypes.func,
+  section: PropTypes.oneOf([
+    'forms',
+    'interactive',
+    'layout',
+    'media',
+    'modals',
+    'popovers',
+  ]),
 };
 
 export const State = ({ render, initialValue }) => {