+++ CHANGELOG.md
@@ -34,6 +34,55 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 #### `Thing` Component
 Component was updated / removed.
 
+## [7.17] Add TextInput
+
+- Added `TextInput` form component for inputs
+
+## [7.16] Updated `Popover`
+
+- Added disableScrollRef to disable an elements scrolling while the popover is visible
+- Added closeOnScrollRef to close the popover when the user starts scrolling the given element
+
+## [7.15] Disabled Menu
+
+- Added `isDisabled` prop to `Menu`
+>>>>>>> @{-1}
+
+## [7.14] Configurable `InputGroup`
+
+- All `InputGroup` components will now fill their parent container using 100% width
+- Updated internal logic of `InputGroup` for setting shrink/grow defaults
+- Added `flexSettings` prop to `InputGroup` to allow for customization of shrink/grow for each input item
+
+## [7.13] Attachment Icon
+
+- Added `AttachmentIcon` to icons
+
+## [7.12] `DateInput`
+
+- Added `DateInput` component.
+
+## [7.11] Disabled MenuItem
+
+- Adds support for `isDisabled` prop in `MenuItem`
+- Changes default z-index for `Menu` to `zindex-popover`
+- Adds `isInModal` prop to `Menu` for bumping up `z-index`
+
+## [7.10] Updated DropdownButton
+
+- DropdownButton now accepts an `isFullWidth`
+- It's background color is now white be default.
+
+## [7.9] Updated Dialog
+
+- Dialog now takes a `height` prop (that's applied as `max-height internally)
+- Dialog now has the option to always render the footer border via `alwaysShowBorder`
+
+## [7.8] Menu + Dialog fixes
+
+- Added `disableFocusTrap` prop to `Dialog`
+- Adjusted `z-index` value of the popover in `Menu`
+
 ## [7.7] Menu component
 
 - Added `Menu` component for building accessible dropdown menus
@@ -414,7 +463,7 @@ MenuButton component was removed. Please use StackedButton instead.
 - Added `customSize` prop to icons to enable arbitrary sizes (we do not
 want to encourage active use of this prop however)
 
-## [4.14] Updated / Added link colors 
+## [4.14] Updated / Added link colors
 
 - Increased link contrast by making default links appear as `blue` (instead of `aqua`)
 Added `linkInverted` property to dictionary and set up inverted link context in base-styles